심사위원 참가,관전,불참 구분

+++ microservices/investment/src/main/java/com/illunex/invest/investment/service/EvaluateService.java
@@ -85,21 +85,20 @@ public class EvaluateService {
         } else {
 
             for (EvaluateReviewItemCategory c : newEvaluate.getTemplate().getReviewItemCategory()) {
-
                 for (EvaluateReviewItem i : c.getReviewItem()) {
                     evaluateReviewItemRepository.deleteAllByReviewItemCategoryIdx(c.getIdx());
                     i.setReviewItemCategory(c);
                 }
-
                 evaluateReviewItemCategoryRepository.deleteAllByReviewItemTemplateIdx(newEvaluate.getTemplate().getIdx());
                 c.setReviewItemTemplate(newEvaluate.getTemplate());
-
             }
 
             for (EvaluateJudge j : newEvaluate.getJudgeList()) {
-                for (EvaluateJudgeScore s : j.getScoreList()) {
-                    evaluateJudgeScoreRepository.deleteAllByJudgeIdx(j.getIdx());
-                    s.setJudge(j);
+                if(j.getDivision().equals("attend")) {
+                    for (EvaluateJudgeScore s : j.getScoreList()) {
+                        evaluateJudgeScoreRepository.deleteAllByJudgeIdx(j.getIdx());
+                        s.setJudge(j);
+                    }
                 }
                 evaluateJudgeRepository.deleteAllByEvaluateIdx(newEvaluate.getIdx());
                 j.setEvaluate(evaluate);
@@ -116,16 +115,20 @@ public class EvaluateService {
             Evaluate editingEvaluate = evaluateRepository.save(evaluate);
 
             for (EvaluateJudge j : newEvaluate.getJudgeList()) {
-                for (EvaluateReviewItemCategory c : editingEvaluate.getTemplate().getReviewItemCategory()) {
-                    for (EvaluateReviewItem i : c.getReviewItem()) {
-                        j.getScoreList().add(EvaluateJudgeScore.builder()
-                            .categoryIdx(c.getIdx())
-                            .reviewItemIdx(i.getIdx())
-                            .score(0)
-                            .judge(j)
-                            .build());
+
+                if(j.getDivision().equals("attend")) {
+                    for (EvaluateReviewItemCategory c : editingEvaluate.getTemplate().getReviewItemCategory()) {
+                        for (EvaluateReviewItem i : c.getReviewItem()) {
+                            j.getScoreList().add(EvaluateJudgeScore.builder()
+                                .categoryIdx(c.getIdx())
+                                .reviewItemIdx(i.getIdx())
+                                .score(0)
+                                .judge(j)
+                                .build());
+                        }
                     }
                 }
+
             }
 
             evaluateRepository.save(editingEvaluate);
@@ -177,10 +180,12 @@ public class EvaluateService {
             return "Invalid Evaluate";
         } else {
             EvaluateJudge judge = mapper.evaluateJudgeDTOToEntity(evaluateReviewDTO.getJudge());
+            int attendCount = 0;
             int completeCount = 0;
 
             for (EvaluateJudge e: evaluate.getJudgeList()) {
-                if (e.getUserIdx().equals(evaluateReviewDTO.getJudge().getUserIdx())) {
+                if ( (e.getUserIdx().equals(evaluateReviewDTO.getJudge().getUserIdx())) && (e.getDivision().equals("attend")) )  {
+                    attendCount++;
                     e.setStatus("confirm");
 
                     for (EvaluateJudgeScore j : e.getScoreList()) {
@@ -218,7 +223,7 @@ public class EvaluateService {
             }
 
 
-            if (completeCount == evaluate.getJudgeList().size()) {
+            if (completeCount == attendCount) {
                 float judgeTotalScore = 0;
 
                 for (EvaluateJudge e: evaluate.getJudgeList()) {
@@ -226,7 +231,7 @@ public class EvaluateService {
                 }
 
                 evaluate.setStatus("complete");
-                evaluate.setAverageScore((int) (judgeTotalScore/3));
+                evaluate.setAverageScore((int) (judgeTotalScore/completeCount));
                 evaluate.setCompleteDate(LocalDateTime.now());
             }
             evaluateRepository.save(evaluate);
